Update Function_ConnectBastionHostRDP.ps1
Changed the Header part for the webrequest.

As this scirpt is written for powershell 5.1 the bearer token must be provided in the request header. This only supports plain text string. While Get-AzAccessToken returns a SecureString this must be resolved bevore using in the authentication header.

On powershell 7.0+ the Authentication parameter of the Invoke-Webrequest can be used which supports SecureStrings by default for authentication.